+++ NailtrestApi/Controllers/CommentsController.cs
@@ -1,26 +1,35 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.IdentityModel.JsonWebTokens;
+using NailtrestApi.Auth.Model;
 using NailtrestApi.Data.Dtos.Comments;
 using NailtrestApi.Data.Entities;
 using NailtrestApi.Data.Repositories;
+using System.Data;
+using System.Security.Claims;
 
 namespace NailtrestApi.Controllers
 {
     [ApiController]
+    [Authorize(Roles = ForumRoles.User)]
     [Route("api/ideas/{ideaId}/comments")]
     public class CommentsController : ControllerBase
     {
         private readonly IIdeasRepository _ideasRepository;
         private readonly ICollectionsRepository _collectionRepository;
         private readonly ICommentsRepository _commentsRepository;
+        private readonly IAuthorizationService _authorizationService;
 
         public CommentsController(
             IIdeasRepository ideasRepository,
             ICollectionsRepository collectionsRepository, 
-            ICommentsRepository commentsRepository)
+            ICommentsRepository commentsRepository,
+            IAuthorizationService authorizationService)
         {
             _ideasRepository = ideasRepository;
             _collectionRepository = collectionsRepository;
             _commentsRepository = commentsRepository;
+            _authorizationService = authorizationService;
         }
 
         [HttpGet]
@@ -60,6 +69,7 @@ namespace NailtrestApi.Controllers
                 Content = createCommentDto.Content,
                 CreatedDate = DateTime.Now,
                 Idea = idea,
+                UserId = User.FindFirstValue(JwtRegisteredClaimNames.Sub)
             };
 
             await _commentsRepository.CreateAsync(comment);
@@ -80,6 +90,13 @@ namespace NailtrestApi.Controllers
 
             var comment = await _commentsRepository.GetAsync(commentId);
 
+            var authResult = await _authorizationService.AuthorizeAsync(User, comment, PolicyNames.ResourceOwner);
+
+            if (!authResult.Succeeded)
+            {
+                return Forbid(); //403
+            }
+
             comment!.Content = updateCommentDto.Content;
 
             await _commentsRepository.UpdateAsync(comment);
@@ -97,6 +114,13 @@ namespace NailtrestApi.Controllers
             }
 
             var comment = await _commentsRepository.GetAsync(commentId);
+
+            var authResult = await _authorizationService.AuthorizeAsync(User, comment, PolicyNames.ResourceOwner);
+
+            if (!authResult.Succeeded)
+            {
+                return Forbid(); //403
+            }
 
             await _commentsRepository.DeleteAsync(comment!);
